this is gonna take a while
