+++ old_src/cli.rs
@@ -3,20 +3,20 @@ use clap::{command, Arg, ArgAction, Command};
 pub fn cli() -> clap::Command {
     let add_subcommand = Command::new("add")
         .alias("new")
-        .about("used to add a new entry to syracuse")
-        .long_about("used to add a new entry to syracuse, you can have aliases for entries, seperate these with spaces")
+        .about("Add a new entry to syracuse\naliases: 'new'")
+        .long_about("This subcommand is used to add a new entry to syracuse, entries are case-insensitive and can have aliases\naliases: 'new'")
         .arg(Arg::new("entry")
                 .index(1)
                 .num_args(1..10)
                 .required(true)
                 .help("entry to add")
-                .long_help("")
+                .long_help("entry to add\ne.g. 'add math-201 analysis' will add an entry titled 'MATH-201' with the alias 'ANALYSIS'")
                 .action(ArgAction::Set)
             );
 
     let list_subcommand = Command::new("list")
-        .aliases(["view", "display", "show"])
-        .about("display all entries")
+        .about("Lists out all entries")
+        .long_about("This subcommand is used to list out all entries stored in syracuse.json")
         .arg(
             Arg::new("full")
                 .short('f')
@@ -25,13 +25,14 @@ pub fn cli() -> clap::Command {
                 .alias("all")
                 .num_args(0)
                 .required(false)
-                .help("also displays the data contained by each entry")
+                .help("prints out the data associated with each entry as well")
                 .action(ArgAction::SetTrue),
         );
 
     let remove_subcommand = Command::new("remove")
         .aliases(["delete", "del"])
-        .about("remove an entry")
+        .about("Removes a single entry\naliases: 'delete', 'del'")
+        .long_about("This subcommand is used to remove a single entry at a time from syracuse.json\naliases: 'delete', 'del'")
         .arg(
             Arg::new("entry")
                 .index(1)
@@ -41,8 +42,9 @@ pub fn cli() -> clap::Command {
         );
 
     let start_subcommand = Command::new("start")
-        .aliases(["s", "r", "run", "go", "launch"])
-        .about("starts the stopwatch for the given entry")
+        .aliases(["s", "r", "run", "go", "launch", "begin"])
+        .about("Starts the daily stopwatch for the given entry")
+        .long_about("This subcommand is used to start counting up the time spent today on the given entry, will progressively update syracuse.json\naliases: 's', 'r', 'run', 'go', 'launch', 'begin'")
         .arg(
             Arg::new("entry")
                 .index(1)
@@ -52,7 +54,8 @@ pub fn cli() -> clap::Command {
         );
 
     let update_subcommand = Command::new("update")
-        .about("updates an entry's data")
+        .about("Manually updates the time of an entry")
+        .long_about("This subcommand is used to manually increase or decrease the time associated with an entry on a given day")
         .arg(
             Arg::new("entry")
                 .index(1)
@@ -60,11 +63,15 @@ pub fn cli() -> clap::Command {
                 .help("entry to update")
                 .action(ArgAction::Set),
         )
+        .subcommand(
+            Command::new("add").alias("a").index(2)
+            Command::new("sub").alias("s").index(2)
+        )
         .arg(
             Arg::new("date")
                 .required(false)
-                .help("updates specified date")
-                .long_help("updates specified date, defaults to today")
+                .help("the targeted date")
+                .long_help("the targeted date, defaults to today")
                 .short('d')
                 .long("date")
                 .action(ArgAction::Set),
@@ -72,15 +79,16 @@ pub fn cli() -> clap::Command {
         .arg(
             Arg::new("hour")
                 .required(false)
-                .help("hours to add or subtract")
-                .short('t')
+                .help("the number of hours to add or subtract")
+                .short('h')
+                .short_alias('t')
                 .long("hour")
                 .action(ArgAction::Set),
         )
         .arg(
             Arg::new("minute")
                 .required(false)
-                .help("minutes to add or subtract")
+                .help("the number of minutes to add or subtract")
                 .short('m')
                 .long("minute")
                 .action(ArgAction::Set),
@@ -88,34 +96,28 @@ pub fn cli() -> clap::Command {
         .arg(
             Arg::new("second")
                 .required(false)
-                .help("secondss to add or subtract")
+                .help("the number of seconds to add or subtract")
                 .short('s')
                 .long("second")
                 .action(ArgAction::Set),
-        )
-        .arg(
-            Arg::new("negative")
-                .required(false)
-                .help("subtract the provided time instead of adding it")
-                .short('n')
-                .long("negative")
-                .action(ArgAction::SetTrue),
         );
 
     let graph_subcommand = Command::new("graph")
-        .alias("export")
-        .about("create a graph")
+        .about("Creates a graph")
+        .long_about("This subcommand is used to graph the entries within a specified time frame")
         .arg(
             Arg::new("all")
+                .help("graphs all entries")
+                .exclusive(true)
                 .short('a')
                 .short('f')
                 .long("all")
                 .alias("full")
-                .exclusive(true)
                 .action(ArgAction::SetTrue),
         )
         .arg(
             Arg::new("single")
+                .help("graphs a single specified entry")
                 .exclusive(true)
                 .short('s')
                 .long("single")
