+++ src/data/mod.rs
@@ -1,2 +1,2 @@
-pub(crate) mod graph;
-pub(crate) mod internal;
+mod internal;
+mod syrtime;
