+++ todo.md
@@ -1,0 +1,3 @@
+modularization of stored entries -> hashed name
+better stability, no mem::transmute for example
+cleaner code
